[wireless tune] using TCP to transmit parameter
Due to the following two reasons, we need to use the TCP protocol to transmit parameters:

UDP's unreliability, which may result in packet loss.
In the implementation of Command,
a single message may call the write interface multiple times,
leading to packet loss and incorrect data parsing.

This is especially problematic when the \n is lost,
as it can cause two messages to merge and result in parsing errors."